+++ src/platform/windows.rs
@@ -77,6 +77,9 @@ pub trait WindowExtWindows {
 
     /// This sets `ICON_BIG`. A good ceiling here is 256x256.
     fn set_taskbar_icon(&self, taskbar_icon: Option<Icon>);
+
+    /// Whether the system theme is currently Windows 10's "Dark Mode".
+    fn is_dark_mode(&self) -> bool;
 }
 
 impl WindowExtWindows for Window {
@@ -93,6 +96,11 @@ impl WindowExtWindows for Window {
     #[inline]
     fn set_taskbar_icon(&self, taskbar_icon: Option<Icon>) {
         self.window.set_taskbar_icon(taskbar_icon)
+    }
+
+    #[inline]
+    fn is_dark_mode(&self) -> bool {
+        self.window.is_dark_mode()
     }
 }