+++ src/platform_impl/windows/window.rs
@@ -34,6 +34,7 @@ use crate::{
     error::{ExternalError, NotSupportedError, OsError as RootOsError},
     monitor::MonitorHandle as RootMonitorHandle,
     platform_impl::platform::{
+        dark_mode::try_dark_mode,
         dpi::{dpi_to_scale_factor, hwnd_dpi},
         drop_handler::FileDropHandler,
         event_loop::{self, EventLoopWindowTarget, DESTROY_MSG_ID, INITIAL_DPI_MSG_ID},
@@ -696,6 +697,11 @@ impl Window {
     pub fn set_ime_position(&self, _logical_spot: LogicalPosition) {
         unimplemented!();
     }
+
+    #[inline]
+    pub fn is_dark_mode(&self) -> bool {
+        self.window_state.lock().is_dark_mode
+    }
 }
 
 impl Drop for Window {
@@ -903,8 +909,19 @@ unsafe fn init<T: 'static>(
     window_flags.set(WindowFlags::VISIBLE, attributes.visible);
     window_flags.set(WindowFlags::MAXIMIZED, attributes.maximized);
 
+    // If the system theme is dark, we need to set the window theme now
+    // before we update the window flags (and possibly show the
+    // window for the first time).
+    let dark_mode = try_dark_mode(real_window.0);
+
     let window_state = {
-        let window_state = WindowState::new(&attributes, window_icon, taskbar_icon, dpi_factor);
+        let window_state = WindowState::new(
+            &attributes,
+            window_icon,
+            taskbar_icon,
+            dpi_factor,
+            dark_mode,
+        );
         let window_state = Arc::new(Mutex::new(window_state));
         WindowState::set_window_flags(window_state.lock(), real_window.0, |f| *f = window_flags);
         window_state